requirements.txt
requirements.txt

Flask==2.3.2
Flask-CORS==4.0.0
Flask-MySQLdb==1.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==2.3.6
MySQLdb-python==1.2.5
